- REST to database working - Scaffolding for DroneDispatcher in place
